Revert "Merge remote-tracking branch 'origin/Merge/Develop' into Merge/Develop"

See merge request jala-university1/cohort-3/oficial-pt-desenvolvimento-de-software-2-iso-124.ga.t2.24.m2/se-o-b/Capstone/demo!18

+++ src/main/java/org/jala/university/domain/repository/repository_card/PasswordCreditCardRepository.java
@@ -3,5 +3,5 @@ package org.jala.university.domain.repository.repository_card;
 import org.jala.university.domain.entity.entity_card.PasswordCreditCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PasswordCreditCardRepository extends JpaRepository<PasswordCreditCard, Long> {
+public interface PasswordCreditCardRepository extends JpaRepository<PasswordCreditCard, Integer> {
 }
